+++ src/app/Scenes/MyCollection/Screens/Artwork/OldMyCollectionArtwork.tsx
@@ -26,7 +26,7 @@ import React from "react"
 import { ScrollView } from "react-native"
 import { graphql, QueryRenderer } from "react-relay"
 import { useTracking } from "react-tracking"
-import { showSubmitToSell } from "../../utils/checkIfSHouldShowSell"
+import { showSubmitToSell } from "../../utils/checkShowSubmitToSell"
 import { MyCollectionArtworkHeaderFragmentContainer } from "./Components/MyCollectionArtworkHeader"
 import { MyCollectionArtworkMetaFragmentContainer } from "./Components/MyCollectionArtworkMeta"
 import { OldWhySell } from "./Components/OldWhySell"
